requirements.txt
# Web server
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles

# Auth & security
authlib
httpx
itsdangerous
bcrypt
slowapi

# Data validation
pydantic

# Environment
python-dotenv

# RAG & LLM
lightrag-hku
openai
langchain-core

# Vector DB
qdrant-client
fastembed

# Embeddings
sentence-transformers
numpy

# Database
asyncpg

# Data ingestion
pypdf
